refactor(core-elasticsearch): simplify storage, server and indexers
