+++ packages/core-elasticsearch/src/index/block.ts
@@ -10,10 +10,6 @@ const logger = app.resolvePlugin<Logger.ILogger>("logger");
 const databaseService = app.resolvePlugin<Database.IDatabaseService>("database");
 
 class BlockIndex extends Index {
-    /**
-     * Index blocks using the specified chunk size.
-     * @return {void}
-     */
     public async index() {
         const { count } = await this.__count();
 
@@ -25,7 +21,7 @@ class BlockIndex extends Index {
             const query = modelQuery
                 .select()
                 .from(modelQuery)
-                .where(modelQuery.timestamp.gte(storage.get("history", "lastBlock")))
+                .where(modelQuery.height.gte(storage.get("lastBlock")))
                 .order(modelQuery.height.asc)
                 .limit(this.chunkSize)
                 .offset(this.chunkSize * i);
@@ -46,8 +42,8 @@ class BlockIndex extends Index {
             try {
                 await client.bulk(this._buildBulkUpsert(rows));
 
-                storage.update("history", {
-                    lastBlock: last(heights),
+                storage.update({
+                    lastBlock: +last(heights),
                 });
             } catch (error) {
                 logger.error(`[Elasticsearch] ${error.message} :exclamation:`);
@@ -55,10 +51,6 @@ class BlockIndex extends Index {
         }
     }
 
-    /**
-     * Register listeners for "block.*" events.
-     * @return {void}
-     */
     public listen() {
         this._registerCreateListener("block.applied");
         // this._registerCreateListener('block.forged')
@@ -66,18 +58,10 @@ class BlockIndex extends Index {
         this._registerDeleteListener("block.reverted");
     }
 
-    /**
-     * Get the document index.
-     * @return {String}
-     */
     public getIndex() {
         return "blocks";
     }
 
-    /**
-     * Get the document type.
-     * @return {String}
-     */
     public getType() {
         return "block";
     }
